Add project directory argument for Ghidra decompiler and improve cleanup process

+++ dataset-gen-ghidra/generate.py
@@ -7,12 +7,13 @@ import signal
 import errno
 import hashlib
 import tempfile as tf
+import shutil
 import pickle
 
 
 from tqdm import tqdm
 from multiprocessing import Pool
-from typing import Iterable, Tuple
+from typing import Iterable, Tuple, Optional
 
 from elftools.elf.elffile import ELFFile
 
@@ -29,6 +30,7 @@ class Runner(object):
         self.verbose = args.verbose
         self.num_threads = args.num_threads
         self.timeout = args.timeout
+        self.project_dir = args.project_dir or self.output_dir
 
         self.env = os.environ.copy()
 
@@ -91,27 +93,36 @@ class Runner(object):
             if e.errno != errno.EEXIST:
                 raise
 
-    def run_decompiler(self, env, path_to_dir, file_name, script, timeout=None):
+    def run_decompiler(self, env, project_parent_dir, file_path, script, timeout=None):
         """Run a decompiler script.
 
         Keyword arguments:
-        file_name -- the binary to be decompiled
+        project_parent_dir -- the directory where the Ghidra project will be created
+        file_path -- the binary file to be decompiled
         env -- an os.environ mapping, useful for passing arguments
         script -- the script file to run
         timeout -- timeout in seconds (default no timeout)
         """
         script_dir = script[:(script.rfind("/"))]
         script_name = script.split("/")[-1]
-        temp_dir = "__".join(file_name.split("/"))
-        orig_file = file_name.split("/")[-1]
+        # Use basename of the file as a project name suffix (unique per file)
+        temp_proj_name = "__".join(os.path.basename(file_path).split("/"))
 
-        ghidracall = [self.ghidra, path_to_dir, temp_dir, '-import', file_name, 
-                      '-postScript', script_name, file_name + ".p", "-scriptPath", script_dir,
-                      '-max-cpu', "1", '-deleteProject']
+        # Create and use a project inside project_parent_dir so Ghidra
+        # does not create project files next to the binary (which might be read-only)
+        ghidracall = [
+            self.ghidra,
+            project_parent_dir,
+            temp_proj_name,
+            '-import',
+            file_path,
+            '-postScript', script_name, file_path + ".p", "-scriptPath", script_dir,
+            '-max-cpu', "1", '-deleteProject']
         if self.verbose:
             print(f"Running {ghidracall}")
+        p = None
         try:
-            p = subprocess.Popen(ghidracall, env=env, start_new_session=True, 
+            p = subprocess.Popen(ghidracall, env=env, start_new_session=True,
                                  stdout=subprocess.PIPE, stderr=subprocess.PIPE)
             stdout, stderr = p.communicate(timeout=timeout)
             if self.verbose:
@@ -119,14 +130,16 @@ class Runner(object):
                 print(stderr.decode())
         except subprocess.TimeoutExpired as e:
             print(f"Timed out while running {ghidracall}")
-            os.killpg(os.getpgid(p.pid), signal.SIGTERM)
-            subprocess.run(f"rm -r {path_to_dir}/__*", shell=True)
+            if p is not None:
+                os.killpg(os.getpgid(p.pid), signal.SIGTERM)
+            # Clean up the project folder which GHIDRA created
+            shutil.rmtree(os.path.join(project_parent_dir, temp_proj_name), ignore_errors=True)
         except subprocess.CalledProcessError as e:
-            subprocess.run(f"rm -r {path_to_dir}/__*", shell=True)
+            shutil.rmtree(os.path.join(project_parent_dir, temp_proj_name), ignore_errors=True)
             if self.verbose:
                 print(e.output.decode())
 
-    def extract_dwarf_var_names(self, filepath:str) -> set:
+    def extract_dwarf_var_names(self, filepath:str) -> Optional[set]:
         """
         Can't figure out how to extract debugging information in GHIDRA's interface, 
         so this function extracts the debugging variable/parameters names to pass into
@@ -200,11 +213,11 @@ class Runner(object):
                         print(f"Collecting debug information")
                     subprocess.check_output(["cp", file_path, orig.name])
 
-                    var_set = self.extract_dwarf_var_names(os.path.join(path, orig.name))
+                    var_set = self.extract_dwarf_var_names(orig.name)
                     if var_set:
-                        pickle_file = os.path.join(path, orig.name) + ".p"
+                        pickle_file = orig.name + ".p"
                         pickle.dump(var_set, open(pickle_file, 'wb'))
-                        self.run_decompiler(new_env, path, os.path.join(path, orig.name), self.COLLECT, timeout=self.timeout)
+                        self.run_decompiler(new_env, self.project_dir, orig.name, self.COLLECT, timeout=self.timeout)
                         os.remove(pickle_file)
                     else:
                         if self.verbose:
@@ -212,11 +225,9 @@ class Runner(object):
                 # Dump trees
                 if self.verbose:
                     print(f"Dumping trees")
-                pickle_file = os.path.join(path, stripped.name) + ".p"
+                pickle_file = stripped.name + ".p"
                 pickle.dump(set(), open(pickle_file, 'wb'))
-                self.run_decompiler(
-                    new_env, path, os.path.join(path, stripped.name), self.DUMP_TREES, timeout=self.timeout
-                )
+                self.run_decompiler(new_env, self.project_dir, stripped.name, self.DUMP_TREES, timeout=self.timeout)
                 os.remove(pickle_file)
 
     def run(self):
@@ -276,6 +287,12 @@ def main():
     )
     parser.add_argument(
         "-o", "--output_dir", metavar="OUTPUT_DIR", help="output directory", required=True,
+    )
+    parser.add_argument(
+        "--project-dir",
+        metavar="PROJECT_DIR",
+        help="parent directory where Ghidra projects are created (defaults to OUTPUT_DIR)",
+        default=None,
     )
     parser.add_argument("--verbose", "-v", action="store_true")
     args = parser.parse_args()
